backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20
websockets==14.2
